Move seek bar handling to `ShakaFrontend`
